Moving to OpenIMU, restructured some stuff.

+++ Quad/Nunchuck.h
@@ -6,18 +6,18 @@
 class Nunchuck
 {  
     public:
-        int accelX;
-        int accelY;
-        int accelZ;
+        int16_t accelX;
+        int16_t accelY;
+        int16_t accelZ;
     
 	void init();
         void init(int calibrationCount);    
         void update();
         
     private:  
-        int accelX0;
-        int accelY0;
-        int accelZ0;
+        int32_t accelX0;
+        int32_t accelY0;
+        int32_t accelZ0;
         
         void _sendByte(byte data, byte location);
 	void _calibrate(int calibrationCount);
